NVIDIA: Refactor `nvidia_symlink()`
The `nvidia_symlink()` currently in GLFS creates absolute symlinks,
which is not great for portability[^1]. This patch addresses that by
pushing and popping directories to create relative symlinks.

It also simplifies some instructions:
1. Parsing the soname has been simplified to get rid of an unreadable
   `sed`. I've elected to use `objdump` over `readelf` because parsing
   strings between square brackets ain't pretty regardless how it's
   done.
2. All instances of `test` have been replaced with `[`, which follows
   the convention in `*LFS-Bootscripts`. It's also more readable imo.
3. Redundant `-f`s have been removed in calls to `ln`. Since we're
   already checking whether something exists before creating the
   symlink, it would never be forcibly created.

Note that I've only tested calls to `nvidia_symlink()` on the targets I
install. I don't expect my changes to break anything, but you might
still want to check that it works for OpenCL, lib32, etc.

[^1]: My DESTDIR install was *mildly* broken. Also in general, making
absolute symlinks for files in the same directory is bad practice.